Move CurrentContext so currentcontext package
