+++ sdk_contrib/testbed/src/test/java/io/opentelemetry/sdk/contrib/trace/testbed/latespanfinish/LateSpanFinishTest.java
@@ -18,7 +18,7 @@ package io.opentelemetry.sdk.contrib.trace.testbed.latespanfinish;
 
 import static com.google.common.truth.Truth.assertThat;
 
-import io.opentelemetry.context.Scope;
+import io.opentelemetry.currentcontext.Scope;
 import io.opentelemetry.exporters.inmemory.InMemoryTracing;
 import io.opentelemetry.sdk.contrib.trace.testbed.TestUtils;
 import io.opentelemetry.sdk.trace.TracerSdkProvider;
