+++ sdk_contrib/testbed/src/test/java/io/opentelemetry/sdk/contrib/trace/testbed/promisepropagation/Promise.java
@@ -16,7 +16,7 @@
 
 package io.opentelemetry.sdk.contrib.trace.testbed.promisepropagation;
 
-import io.opentelemetry.context.Scope;
+import io.opentelemetry.currentcontext.Scope;
 import io.opentelemetry.trace.Span;
 import io.opentelemetry.trace.Tracer;
 import java.util.ArrayList;
